[update] composite 패턴 적용
